+++ elasticapm/contrib/grpc/async_server_interceptor.py
@@ -28,76 +28,18 @@
 #  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 #  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+import inspect
+
 import grpc
-import wrapt
 
 import elasticapm
-from elasticapm.conf.constants import OUTCOME
-from elasticapm.contrib.grpc.utils import STATUS_TO_OUTCOME
-from elasticapm.utils.disttracing import TraceParent
+from elasticapm.contrib.grpc.server_interceptor import _ServicerContextWrapper, _wrap_rpc_behavior, get_trace_parent
 
 
-# from https://github.com/grpc/grpc/issues/18191
-def _wrap_rpc_behavior(handler, continuation):
-    if handler is None:
-        return None
-
-    if handler.request_streaming and handler.response_streaming:
-        behavior_fn = handler.stream_stream
-        handler_factory = grpc.stream_stream_rpc_method_handler
-    elif handler.request_streaming and not handler.response_streaming:
-        behavior_fn = handler.stream_unary
-        handler_factory = grpc.stream_unary_rpc_method_handler
-    elif not handler.request_streaming and handler.response_streaming:
-        behavior_fn = handler.unary_stream
-        handler_factory = grpc.unary_stream_rpc_method_handler
-    else:
-        behavior_fn = handler.unary_unary
-        handler_factory = grpc.unary_unary_rpc_method_handler
-
-    return handler_factory(
-        continuation(behavior_fn, handler.request_streaming, handler.response_streaming),
-        request_deserializer=handler.request_deserializer,
-        response_serializer=handler.response_serializer,
-    )
-
-
-def get_trace_parent(handler_call_details):
-    traceparent, tracestate = None, None
-    for metadata in handler_call_details.invocation_metadata:
-        if metadata.key == "traceparent":
-            traceparent = metadata.value
-        elif metadata.key == "tracestate":
-            tracestate = metadata.key
-    if traceparent:
-        return TraceParent.from_string(traceparent, tracestate)
-    else:
-        return None
-
-
-class _ServicerContextWrapper(wrapt.ObjectProxy):
-    def __init__(self, wrapped, transaction):
-        self._self_transaction = transaction
-        super().__init__(wrapped)
-
-    def abort(self, code, details):
-        transaction = self._self_transaction
-        if transaction:
-            transaction.set_failure()
-        return self.__wrapped__.abort(code, details)
-
-    def set_code(self, code):
-        transaction = self._self_transaction
-        if transaction:
-            outcome = STATUS_TO_OUTCOME.get(code, OUTCOME.SUCCESS)
-            transaction.set_success() if outcome is OUTCOME.SUCCESS else transaction.set_failure()
-        return self.__wrapped__.set_code(code)
-
-
-class _ServerInterceptor(grpc.ServerInterceptor):
-    def intercept_service(self, continuation, handler_call_details):
+class _AsyncServerInterceptor(grpc.aio.ServerInterceptor):
+    async def intercept_service(self, continuation, handler_call_details):
         def transaction_wrapper(behavior, request_streaming, response_streaming):
-            def _interceptor(request_or_iterator, context):
+            async def _interceptor(request_or_iterator, context):
                 if request_streaming or response_streaming:  # only unary-unary is supported
                     return behavior(request_or_iterator, context)
                 tp = get_trace_parent(handler_call_details)
@@ -105,6 +47,11 @@ class _ServerInterceptor(grpc.ServerInterceptor):
                 transaction = client.begin_transaction("request", trace_parent=tp)
                 try:
                     result = behavior(request_or_iterator, _ServicerContextWrapper(context, transaction))
+
+                    # This is so we can support both sync and async rpc functions
+                    if inspect.isawaitable(result):
+                        result = await result
+
                     if transaction and not transaction.outcome:
                         transaction.set_success()
                     return result
@@ -118,4 +65,4 @@ class _ServerInterceptor(grpc.ServerInterceptor):
 
             return _interceptor
 
-        return _wrap_rpc_behavior(continuation(handler_call_details), transaction_wrapper)
+        return _wrap_rpc_behavior(await continuation(handler_call_details), transaction_wrapper)
